frog-auto-grader/api-gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
firebase-admin==6.3.0
authlib==1.2.1
httpx==0.25.1
pydantic==2.5.0
pydantic-settings==2.1.0